Raffles, Chests, Dialog handling
Implemented logic to open raffles in the balanced arena.
Implemented logic to open veltain treasure chests.
Implemented an enum Event Type for handling timed activities
Updated the dialog packet handler and implemented helper methods to handle various dialog messages from npcs.
Implemented a method to save timed information in the bot folder.

+++ Objects/Inventory.cs
@@ -76,7 +76,7 @@ internal sealed class Inventory : IEnumerable<Item>, IEnumerable
         int totalCount = 0;
         for (int index = 0; index < MAX_ITEMS; ++index)
         {
-            if (item[index] != null && (int)item[index].Sprite - CONSTANTS.ITEM_SPRITE_OFFSET == (int)sprite)
+            if (item[index] != null && item[index].Sprite - CONSTANTS.ITEM_SPRITE_OFFSET == sprite)
             {
                 totalCount += includeStack ? item[index].Quantity : 1;
             }
